pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mesh_voxelisation"
version = "0.1.0"
description = "Voxelise closed triangular-polygon meshes (Python port of A. Aitkenhead's MATLAB code)."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "BSD-3-Clause" }
authors = [
    { name = "Adam H. Aitkenhead (original MATLAB)" },
    { name = "Python port" },
]
dependencies = [
    "numpy>=1.20",
]

[project.optional-dependencies]
plot = ["matplotlib>=3.5"]
tiff = ["Pillow>=9.0"]
dev = ["pytest>=7.0"]

[tool.setuptools]
packages = ["mesh_voxelisation"]
